split planet.go file to avoid package naming conflicts
